chore(list): move list result to its own entity and export it
